--- a/docs/.vitepress/configs/version.ts
+++ b/docs/.vitepress/configs/version.ts
@@ -1,1 +1,1 @@
-export const version: string = '2.15.0-preview1'
+export const version: string = '2.15.0-preview2'