--- a/src/Gridify.Elasticsearch/Gridify.Elasticsearch.csproj
+++ b/src/Gridify.Elasticsearch/Gridify.Elasticsearch.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
    <PropertyGroup>
       <PackageId>Gridify.Elasticsearch</PackageId>
-      <Version>2.15.0-preview1</Version>
+      <Version>2.15.0-preview2</Version>
       <Authors>Alireza Sabouri; Dzmitry Koush</Authors>
       <PackageDescription>Gridify (Elasticsearch), Easy way to apply Filtering, Sorting, and Pagination using text-based data.</PackageDescription>
       <RepositoryUrl>https://github.com/alirezanet/Gridify</RepositoryUrl>