--- a/src/Gridify/Gridify.csproj
+++ b/src/Gridify/Gridify.csproj
@@ -2,7 +2,7 @@
 
    <PropertyGroup>
       <PackageId>Gridify</PackageId>
-      <Version>2.15.0-preview1</Version>
+      <Version>2.15.0-preview2</Version>
       <Authors>Alireza Sabouri</Authors>
       <Company>TuxTeam</Company>
       <PackageDescription>Gridify, Easy and optimized way to apply Filtering, Sorting, and Pagination using text-based data.</PackageDescription>