--- a/src/Gridify/Syntax/ISyntaxNode.cs
+++ b/src/Gridify/Syntax/ISyntaxNode.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace Gridify.Syntax;
+
+public interface ISyntaxNode
+{
+   public SyntaxKind Kind { get; }
+   public IEnumerable<ISyntaxNode> GetChildren();
+}