--- a/src/Gridify/Syntax/SyntaxNode.cs
+++ b/src/Gridify/Syntax/SyntaxNode.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Gridify.Syntax;
-
-public abstract class SyntaxNode
-{
-   public abstract SyntaxKind Kind { get; }
-   public abstract IEnumerable<SyntaxNode> GetChildren();
-}