… benchmark/NativeLinqComparisionBenchmark.cs
… benchmark/NativeLinqComparisonBenchmark.cs
… benchmark/Program.cs
… src/Gridify/Syntax/OperatorManager.cs
… src/Gridify/Syntax/ParenthesizedExpressionSyntax.cs
… src/Gridify/Syntax/Parser.cs
… src/Gridify/Syntax/SimpleTypeHelper.cs
… src/Gridify/Syntax/SyntaxToken.cs
… src/Gridify/Syntax/ValueExpressionSyntax.cs
… test/Gridify.Tests/IssueTests/Issue168Tests.cs