BenchmarkDotNet v0.13.10, Windows 11 (10.0.22631.3737/23H2/2023Update/SunValley3)
AMD Ryzen 7 7800X3D, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.206
  [Host]     : .NET 8.0.6 (8.0.624.26715), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.6 (8.0.624.26715), X64 RyuJIT AVX2


| Method                | Mean     | Error    | StdDev   | Ratio | RatioSD | Allocated | Alloc Ratio |
|---------------------- |---------:|---------:|---------:|------:|--------:|----------:|------------:|
| Gridify               | 591.8 us |  5.41 us |  5.07 us |  0.92 |    0.04 |  35.65 KB |        1.09 |
| Gridify_WithoutMapper | 603.4 us |  4.39 us |  4.10 us |  0.94 |    0.04 |  40.81 KB |        1.25 |
| Native_LINQ           | 636.1 us | 11.31 us | 18.90 us |  1.00 |    0.00 |  32.63 KB |        1.00 |